Proposed refactor on service
