+++ src/Pages/ChannelRequests.razor
@@ -5,6 +5,7 @@
 @using Humanizer
 @using NBitcoin
 @using FundsManager.Jobs
+@using FundsManager.Services.Interfaces
 @attribute [Authorize(Roles = "FinanceManager, Superadmin, NodeManager")]
 
 <PageTitle>Channel Operation Requests</PageTitle>
