+++ src/Pages/Shared/_Layout.cshtml
@@ -26,7 +26,6 @@
 </head>
 <body>
     <div class="container">
-        <partial name="_CookieConsentPartial" optional />
         <main role="main" class="pb-1">
             @RenderBody()
         </main>
