+++ src/Pages/Wallets.razor
@@ -1,5 +1,6 @@
 @page "/wallets"
 @using System.Security.Claims
+@using FundsManager.Services.Interfaces
 @using Humanizer
 @using NBitcoin
 @using NBXplorer.DerivationStrategy
